Refactor + hospital queue in main
