+++ src/FibHeap2.java
@@ -1,279 +1,224 @@
-import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
 
 public class FibHeap2 {
-    private FibNode2 minNode;
-    private int size;
-    private FibNode2 last;
-    private FibNode2 root;
+    
+    // Pointer to the head and minimum node in the root list
+    private FibNode2 rootList, minNode;
 
-    public FibHeap2() {
-        minNode = null;
-        last = null;
-        size = 0;
-        root = null;
-    }
-    
-    public void insert(FibNode2 node) {
-        if (node == null) {
-            return;
-        }
-        
-        node.next = node;
-    	node.prev = node;
-       
-        addToRootList(node);
-        if (minNode == null || node.priority < minNode.priority) {
-            minNode = node;
-        }
-        size++;
-        last = node;
-    }
+    // Maintain total node count in full Fibonacci heap
+    private int totalFibNode2s; 
 
-    public FibNode2 getLast() {
-        return last;
-    }
-    
-    public FibNode2 getRoot() {
-        return root;
-    }
-
-    public boolean isEmpty() {
-        if(minNode == null) {
-        	return true;
-        }
-        return false;
-    }
-    
-    public void clear() {
-    	root = null;
-        minNode = null;
-        size = 0;
-    }
-    
-    public int size() {
-        return size;
-    }
-    
+    // Return min node in O(1) time
     public FibNode2 peek() {
         return minNode;
     }
 
-    public FibNode2 getMin() {
-        if (isEmpty()) {
-            return null;
-        }
-        FibNode2 first = minNode, child = minNode.child, anchor = minNode.child;
-        if (child != null && minNode != null) {
-            do {
-            	FibNode2 nextChild = child.next;
-                addToRootList(child);
-                child.parent = null;
-                child = nextChild;
-            } while (child != anchor);
-        }
-        removeMin();
-        return first;
-    }
-
-    public FibNode2 findNode(String name) {
-        if (isEmpty()) {
-            return null; 
-        } else {
-            return findNode(minNode, name);
-        }
-    }
-
-    public FibNode2 removeNode(FibNode2 node) {
-        if (node == null) {
-            return null; 
-        }
-        if (node == minNode) {
-            removeMin();
-        } else {
-            cut(node, node.parent);
-            cascadingCut(node);
-        }
-        size--;
-        return node;
-    }
-
-    private void removeMin() {
-    	removeFromRootList(minNode);
-        if (minNode.next == minNode) {
-            minNode = null;
-            root = null;
-        } else {
-            minNode = minNode.next;
-            consolidate();
-        }
-        size--;
-    }
-    
-
-    private FibNode2 findNode(FibNode2 start, String name) {
-        FibNode2 current = start;
-        while (current != null) {
-            if (current.name.equals(name)) {
-                return current;
-            }
-    
-            FibNode2 foundInChild = findNode(current.child, name);
-            if (foundInChild != null) {
-                return foundInChild;
-            }
-    
-            current = current.next;
-            if (current == start) {
-                break;
-            }
-        }
-        return null;
-    }  
-
-    private void consolidate() {
-        int maxDegree = (int) Math.floor(Math.log(size) / Math.log(2)) + 1;
-    
-        FibNode2[] degreeTable = new FibNode2[maxDegree];
-    
-        // Loop will run at least once when roots is empty
-        ArrayList<FibNode2> roots = new ArrayList<>();
-        for (FibNode2 current = minNode; roots.isEmpty() || current != minNode; current = current.next) {
-            roots.add(current);
-        }
-    
-        for (FibNode2 root : roots) {
-            FibNode2 current = root;
-            int degree = current.degree;
-    
-            while (degreeTable[degree] != null) {
-                FibNode2 other = degreeTable[degree];
-                if (current.priority > other.priority) {
-                    FibNode2 temp = current;
-                    current = other;
-                    other = temp;
-                }
-    
-                link(other, current);
-                degreeTable[degree] = null;
-                degree++;
-            }
-    
-            degreeTable[degree] = current;
-            
-        }
-        // recheck for min node
-        minNode = null;
-        for (FibNode2 node : degreeTable) {
-            if (node != null) {
-                if (minNode == null || node.priority < minNode.priority) {
-                    minNode = node;
+    // Extract (delete) the min node from the heap in O(log n) time
+    public FibNode2 dequeue() {
+        FibNode2 tempNode = minNode;
+        if (tempNode != null) {
+            if (tempNode.child != null) {
+                // Attach child nodes to the root list
+                List<FibNode2> children = iterate(tempNode.child);
+                for (FibNode2 child : children) {
+                    mergeWithRootList(child);
+                    child.parent = null;
                 }
             }
-        }
-    }
-
-    private void link(FibNode2 child, FibNode2 parent) {
-        child.prev.next = child.next;
-        child.next.prev = child.prev;
-    
-        child.prev = child;
-        child.next = child;
-        parent.degree++;
-    
-        if (parent.child == null) {
-            parent.child = child;
-        } else {
-            child.next = parent.child;
-            child.prev = parent.child.prev;
-            parent.child.prev.next = child;
-            parent.child.prev = child;
-        }
-        child.marked = false;
-        child.parent = parent;
-    }
-
-    private void addToRootList(FibNode2 newNode) {
-        if (root == null) {
-            root = newNode;
-        } else {
-            newNode.next = root;
-            newNode.prev = root.prev;
-            root.prev.next = newNode;
-            root.prev = newNode;
-        }
-    }
-    
-    private void removeFromRootList(FibNode2 node) {
-    	if (node == root) {
-    		root = node.next;
-    	}
-		node.prev.next = node.next;
-		node.next.prev = node.prev;
-    }
-
-    private void cut(FibNode2 node, FibNode2 parent) {
-        if (parent != null) {
-            if (node.next == node) {
-                parent.child = null;
+            removeFromRootList(tempNode);
+            // Set new min node in heap
+            if (tempNode == tempNode.right) {
+                minNode = rootList = null;
             } else {
-                node.next.prev = node.prev;
-                node.prev.next = node.next;
-                if (parent.child == node) {
-                    parent.child = node.next;
-                }
+                minNode = tempNode.right;
+                consolidate();
             }
-            parent.degree--;
-            addToRootList(node);
-            node.parent = null;
-            node.marked = false;
-            if (node.priority < minNode.priority) {
-                minNode = node;
-            }
+            totalFibNode2s--;
+        }
+        return tempNode;
+    }
+
+    // Insert new node into the unordered root list in O(1) time
+    // Returns the node so that it can be used for decreasepriority later
+    public FibNode2 insert(String name, int priority) {
+        FibNode2 newNode = new FibNode2(name, priority);
+        newNode.left = newNode.right = newNode;
+        mergeWithRootList(newNode);
+        if (minNode == null || newNode.priority < minNode.priority) {
+            minNode = newNode;
+        }
+        totalFibNode2s++;
+        return newNode;
+    }
+
+    // Modify the priority of some node in the heap in O(1) time
+    public void decreasepriority(FibNode2 node, int newPriority) {
+        if (newPriority > node.priority) {
+            return;
+        }
+        node.priority = newPriority;
+        FibNode2 parent = node.parent;
+        if (parent != null && node.priority < parent.priority) {
+            cut(node, parent);
+            cascadingCut(parent);
+        }
+        if (node.priority < minNode.priority) {
+            minNode = node;
         }
     }
-    
+
+    // Merge two Fibonacci heaps in O(1) time by concatenating the root lists
+    // The root of the new root list becomes equal to the first list, and the second
+    // list is simply appended to the end (then the proper min node is determined)
+    public FibHeap2 merge(FibHeap2 h2) {
+    	FibHeap2 H = new FibHeap2();
+        H.rootList = rootList;
+        H.minNode = minNode;
+        // Fix pointers when merging the two heaps
+        FibNode2 last = h2.rootList.left;
+        h2.rootList.left = H.rootList.left;
+        H.rootList.left.right = h2.rootList;
+        H.rootList.left = last;
+        H.rootList.left.right = H.rootList;
+        // Update min node if needed
+        if (h2.minNode.priority < H.minNode.priority) {
+            H.minNode = h2.minNode;
+        }
+        // Update total nodes
+        H.totalFibNode2s = totalFibNode2s + h2.totalFibNode2s;
+        return H;
+    }
+
+    // Function to iterate through a doubly linked list
+    private List<FibNode2> iterate(FibNode2 head) {
+        List<FibNode2> nodes = new LinkedList<>();
+        FibNode2 node = head;
+        FibNode2 stop = head;
+        boolean flag = false;
+
+        do {
+            if (node == stop && flag) {
+                break;
+            } else if (node == stop) {
+                flag = true;
+            }
+            nodes.add(node);
+            node = node.right;
+        } while (true);
+
+        return nodes;
+    }
+
+    // If a child node becomes smaller than its parent node, we
+    // cut this child node off and bring it up to the root list
+    private void cut(FibNode2 node1, FibNode2 node2) {
+        removeFromChildList(node2, node1);
+        node2.degree--;
+        mergeWithRootList(node1);
+        node1.parent = null;
+        node1.mark = false;
+    }
+
+    // Cascading cut of parent node to obtain good time bounds
     private void cascadingCut(FibNode2 node) {
         FibNode2 parent = node.parent;
         if (parent != null) {
-            if (!node.marked) {
-                node.marked = true;
+            if (!node.mark) {
+                node.mark = true;
             } else {
                 cut(node, parent);
                 cascadingCut(parent);
             }
         }
     }
-    
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("FibonacciHeap [\n");
-        
-        if (minNode != null) {
-            sb.append("  Min Node: ").append(minNode.priority).append("\n");
-        } else {
-            sb.append("  Min Node: null\n");
-        }
 
-        sb.append("  Size: ").append(size).append("\n");
-
-        if (root != null) {
-            sb.append("  Root List: [");
-            FibNode2 current = root;
-            do {
-                sb.append(current.priority);
-                current = current.next;
-                if (current != root) {
-                    sb.append(", ");
+    // Combine root nodes of equal degree to consolidate the heap
+    // by creating a list of unordered binomial trees
+    private void consolidate() {
+        int maxDegree = (int) Math.floor(Math.log(totalFibNode2s) / Math.log(2)) + 1;
+        FibNode2[] nodeList = new FibNode2[maxDegree];
+        List<FibNode2> nodes = iterate(rootList);
+        for (FibNode2 curNode : nodes) {
+            int degree = curNode.degree;
+            while (nodeList[degree] != null) {
+                FibNode2 tempNode = nodeList[degree];
+                if (curNode.priority > tempNode.priority) {
+                    FibNode2 temp = curNode;
+                    curNode = tempNode;
+                    tempNode = temp;
                 }
-            } while (current != root);
-            sb.append("]\n");
-        } else {
-            sb.append("  Root List: empty\n");
+                heapLink(tempNode, curNode);
+                nodeList[degree] = null;
+                degree++;
+            }
+            nodeList[degree] = curNode;
         }
-
-        sb.append("]");
-
-        return sb.toString();
+        // Find new min node - no need to reconstruct new root list below
+        // because the root list was iteratively changing as we were moving
+        // nodes around in the above loop
+        for (FibNode2 node : nodeList) {
+            if (node != null) {
+                if (node.priority < minNode.priority) {
+                    minNode = node;
+                }
+            }
+        }
     }
 
+    // Actual linking of one node to another in the root list
+    // while also updating the child linked list
+    private void heapLink(FibNode2 node1, FibNode2 node2) {
+        removeFromRootList(node1);
+        node1.left = node1.right = node1;
+        mergeWithChildList(node2, node1);
+        node2.degree++;
+        node1.parent = node2;
+        node1.mark = false;
+    }
+
+    // Merge a node with the doubly linked root list
+    private void mergeWithRootList(FibNode2 node) {
+        if (rootList == null) {
+            rootList = node;
+        } else {
+            node.right = rootList.right;
+            node.left = rootList;
+            rootList.right.left = node;
+            rootList.right = node;
+        }
+    }
+
+    // Merge a node with the doubly linked child list of a root node
+    private void mergeWithChildList(FibNode2 parent, FibNode2 node) {
+        if (parent.child == null) {
+            parent.child = node;
+        } else {
+            node.right = parent.child.right;
+            node.left = parent.child;
+            parent.child.right.left = node;
+            parent.child.right = node;
+        }
+    }
+
+    // Remove a node from the doubly linked root list
+    private void removeFromRootList(FibNode2 node) {
+        if (node == rootList) {
+            rootList = node.right;
+        }
+        node.left.right = node.right;
+        node.right.left = node.left;
+    }
+
+    // Remove a node from the doubly linked child list
+    private void removeFromChildList(FibNode2 parent, FibNode2 node) {
+        if (parent.child == parent.child.right) {
+            parent.child = null;
+        } else if (parent.child == node) {
+            parent.child = node.right;
+            node.right.parent = parent;
+        }
+        node.left.right = node.right;
+        node.right.left = node.left;
+    }
 }
