+++ src/HospitalQueue.java
@@ -1,6 +1,7 @@
 public class HospitalQueue {
 
     private FibHeap2 fibHeap;
+    private int numPatients = 0;
 
     /**
      * Constructor
@@ -12,29 +13,30 @@ public class HospitalQueue {
     /**
      * Adds a new patient to the queue
      */
-    public void insert(String name, int priority) {
-        FibNode2 newNode = new FibNode2(name, priority);
-        fibHeap.insert(newNode);
+    public void addPatient(String name, int priority) {
+        numPatients++;
+        fibHeap.insert(name, priority);
     }
 
     /**
      * Changes the priority of the patient
      */
-    public void changePriority(String name, int newPriority) {
-        fibHeap.removeNode(this.checkPatient(name));
-        this.insert(name, newPriority);
+    public void decreasePriority(FibNode2 patient, int newPriority) {
+        fibHeap.decreasepriority(patient, newPriority);;
     }
 
     /**
-     * Shows the first patient in the queue's name and priority
+     * Shows the first patient in the queue's name
      */
     public String firstPatient() {
-        String patient = fibHeap.peek().name;
-        int priority = fibHeap.peek().priority;
-        return ("[ " + patient + " " + priority + " ]");
+        return fibHeap.peek().name;
     }
 
-    public int highestPriority() {
+    /**
+     * Shows the priority of the patient first in line
+     * @return
+     */
+    public int firstPriority() {
         return fibHeap.peek().priority;
     }
 
@@ -42,55 +44,39 @@ public class HospitalQueue {
      * Dequeues the first patient in the queue
      */
     public FibNode2 dequeue() {
-        return fibHeap.getMin();
+        if (!this.isEmpty()) {
+           numPatients--; 
+           return fibHeap.dequeue();
+        }
+        return null;
     }
 
     /**
-     * Dequeues the patient by name
+     * Returns but does not remove the first patient in the queue
      */
-    public FibNode2 dequeue(String name) {
-        return fibHeap.removeNode(this.checkPatient(name));
-    }
-
-    /**
-     * Gets the first patient in the queue and does not remove them
-     */
-    public FibNode2 checkFirstPatient() {
+    public FibNode2 peek() {
         return fibHeap.peek();
-    }
-    
-    /**
-     * Gets the last patient in the queue and does not remove them
-     */
-    public FibNode2 checkLastPatient() {
-        return fibHeap.getLast();
-    }
-
-    /**
-     * Gets the patient of a specific name and does not remove them
-     */
-    public FibNode2 checkPatient(String name) {
-        return fibHeap.findNode(name);
     }
 
     /**
      * Checks if the number of patients is 0
      */
     public boolean isEmpty() {
-        return fibHeap.isEmpty();
+        return numPatients == 0;
     }
 
     /**
      * Empties the queue
      */
     public void emptyQueue() {
-        fibHeap.clear();
+        numPatients = 0;
+        fibHeap = new FibHeap2();
     }
 
     /**
      * gets the number of patients
      */
     public int numPatients() {
-        return fibHeap.size();
+        return numPatients;
     }
 }
